#17 Renamed SyntaxTreeBuilder to ModelBuilder
