+++ cml-compiler/cml-language/src/main/java/cml/language/interfaces/builder/BeforeRule.java
@@ -1,4 +1,4 @@
-package cml.language.interfaces.tree;
+package cml.language.interfaces.builder;
 
 @interface BeforeRule
 {
